feat(identity-duck): refactor identity duck
