+++ src/common/identity/types.js
@@ -1,0 +1,34 @@
+export const identityTypes = {
+	IDENTITY_REPOSITORIES_LOAD: 'identity/repositories/LOAD',
+	IDENTITY_REPOSITORIES_SET: 'identity/repositories/SET',
+	IDENTITY_REPOSITORIES_UPDATE_REMOTE: 'identity/repositories/UPDATE_REMOTE',
+	IDENTITY_ID_ATTRIBUTE_TYPES_LOAD: 'identity/id-atribute_types/LOAD',
+	IDENTITY_ID_ATTRIBUTE_TYPES_SET: 'identity/id-atribute_types/SET',
+	IDENTITY_ID_ATTRIBUTE_TYPES_UPDATE_REMOTE: 'identity/id-atribute_types/UPDATE_REMOTE',
+	IDENTITY_UI_SCHEMAS_LOAD: 'identity/ui-schemas/LOAD',
+	IDENTITY_UI_SCHEMAS_SET: 'identity/ui-schemas/SET',
+	IDENTITY_UI_SCHEMAS_UPDATE_REMOTE: 'identity/ui-schemas/UPDATE_REMOTE',
+	IDENTITY_DOCUMENTS_LOAD: 'identity/documents/LOAD',
+	IDENTITY_DOCUMENTS_SET: 'identity/documents/SET',
+	IDENTITY_DOCUMENT_ADD: 'identity/documents/ADD',
+	IDENTITY_DOCUMENT_UPDATE: 'identity/documents/ADD',
+	IDENTITY_DOCUMENTS_DELETE: 'identity/documents/DELETE_ONE',
+	IDENTITY_DOCUMENT_DELETE: 'identity/documents/DELETE',
+	IDENTITY_ATTRIBUTES_LOAD: 'identity/attributes/LOAD',
+	IDENTITY_ATTRIBUTES_SET: 'identity/attributes/SET',
+	IDENTITY_ATTRIBUTE_ADD: 'identity/attribute/ADD',
+	IDENTITY_ATTRIBUTE_REMOVE: 'identity/attribute/REMOVE',
+	IDENTITY_ATTRIBUTE_CREATE: 'identity/attribute/CREATE',
+	IDENTITY_ATTRIBUTE_UPDATE: 'identity/attribute/UPDATE',
+	IDENTITY_ATTRIBUTE_EDIT: 'identity/attribute/EDIT',
+	IDENTITY_ATTRIBUTES_DELETE: 'identity/attributes/DELETE',
+	IDENTITY_ATTRIBUTE_DELETE: 'identity/attributes/DELETE_ONE',
+	IDENTITY_ATTRIBUTE_DOCUMENTS_LOAD: 'identity/attribute_documents/LOAD',
+	IDENTITY_ATTRIBUTE_DOCUMENTS_SET: 'identity/attribute_documents/SET',
+	IDENTITY_ATTRIBUTE_DOCUMENTS_DELETE: 'identity/attribute_documents/DELETE',
+	IDENTITY_DOCUMENT_REMOVE: 'identity/documents/REMOVE',
+	IDENTITY_UNLOCK: 'identity/UNLOCK',
+	IDENTITY_LOCK: 'identity/LOCK'
+};
+
+export default identityTypes;
